# docs/openapi3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Konveyor Hub API",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.3.z"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/applications/{id}/tags": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "List tag references.",
        "description": "List tag references.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Add tag association.",
        "description": "Ensure tag is associated with the application.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Tag data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Ref"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "tag"
      },
      "patch": {
        "tags": [
          "applications"
        ],
        "summary": "Replace tag associations.",
        "description": "Replace tag associations.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Source",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Tag references",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.TagRef"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "tags"
      }
    },
    "/applications/{id}/tags/{sid}": {
      "delete": {
        "tags": [
          "applications"
        ],
        "summary": "Delete tag association.",
        "description": "Ensure tag is not associated with the application.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sid",
            "in": "path",
            "description": "Tag ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      }
    },
...
    "/batch/tags": {
      "post": {
        "tags": [
          "batch",
          "tags"
        ],
        "summary": "Batch-create Tags.",
        "description": "Batch-create Tags.",
        "requestBody": {
          "description": "Tags data",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Tag"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Tag"
                  }
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "tags"
      }
    },
...
    "/tagcategories/{id}/tags": {
      "get": {
        "tags": [
          "tagcategories"
        ],
        "summary": "List the tags in the tag category.",
        "description": "List the tags in the tag category.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Tag Category ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Optional tag name filter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Tag"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tags": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "List all tags.",
        "description": "List all tags.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Tag"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tags"
        ],
        "summary": "Create a tag.",
        "description": "Create a tag.",
        "requestBody": {
          "description": "Tag data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Tag"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Tag"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "tag"
      }
    },
    "/tags/{id}": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "Get a tag by ID.",
        "description": "Get a tag by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Tag ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Tag"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tags"
        ],
        "summary": "Update a tag.",
        "description": "Update a tag.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Tag ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Tag data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Tag"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "tag"
      },
      "delete": {
        "tags": [
          "tags"
        ],
        "summary": "Delete a tag.",
        "description": "Delete a tag.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Tag ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "github_com_konveyor_tackle2-hub_api.Ref": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "github_com_konveyor_tackle2-hub_api.Tag": {
        "required": [
          "category",
          "name"
        ],
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
          },
          "createTime": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          }
        }
      },
...
      "github_com_konveyor_tackle2-hub_internal_api.Ref": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "github_com_konveyor_tackle2-hub_internal_api.Tag": {
        "required": [
          "category",
          "name"
        ],
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
          },
          "createTime": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          }
        }
      },
      "github_com_konveyor_tackle2-hub_internal_api.TagRef": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "virtual": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}